# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CreateEventResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"hash": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"hash"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RequestGroup": {
				"description": "A group is a single organization that is an end customer of a vendor app",
				"properties": {
					"id": {
						"type": "string",
						"description": "The id of this group in the vendor app's data model"
					},
					"name": {
						"type": "string",
						"description": "A human-readable name"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EventFields": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"type": "string"
				}
			},
			"RequestActor": {
				"description": "An actor is the person or identity (like an API token) that performed the action",
				"properties": {
					"id": {
						"type": "string",
						"description": "The id of this actor in the vendor app's data model"
					},
					"name": {
						"type": "string",
						"description": "A human-readable name"
					},
					"href": {
						"type": "string",
						"description": "A url to view this actor in the vendor app.\nCan be referenced in Retraced [Display Templates](https://boxyhq.com/docs/retraced/advanced/display-templates)\nto create an interactive embedded viewer experience."
					},
					"fields": {
						"$ref": "#/components/schemas/EventFields"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"RequestTarget": {
				"description": "A target is the object upon which the action is performed",
				"properties": {
					"id": {
						"type": "string",
						"description": "The id of this target in the vendor app's data model"
					},
					"name": {
						"type": "string",
						"description": "A human-readable name"
					},
					"href": {
						"type": "string",
						"description": "A url to view this target in the vendor app.\nCan be referenced in Retraced\nto create an interactive embedded viewer experience."
					},
					"type": {
						"type": "string",
						"description": "Identifies the type"
					},
					"fields": {
						"$ref": "#/components/schemas/EventFields"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"CreateEventRequest": {
				"properties": {
					"action": {
						"type": "string",
						"description": "The action that occured e.g. `user.login` or `spreadsheet.create`"
					},
					"crud": {
						"type": "string",
						"description": "Denotes whether this is a \"Create\", \"Read\", \"Update\", or \"Delete\" event."
					},
					"group": {
						"$ref": "#/components/schemas/RequestGroup"
					},
					"created": {
						"type": "string",
						"format": "date-time",
						"description": "ISO8601 date string representing when this event occurent. `created` will be tracked in addtion to `received`"
					},
					"actor": {
						"$ref": "#/components/schemas/RequestActor"
					},
					"target": {
						"$ref": "#/components/schemas/RequestTarget"
					},
					"source_ip": {
						"type": "string",
						"description": "The source IP address from which the event was initiated"
					},
					"description": {
						"type": "string",
						"description": "A human-readable description of the event"
					},
					"is_anonymous": {
						"type": "boolean",
						"description": "Denotes whether this event was anonymous. Must be `true` if `actor` is absent"
					},
					"is_failure": {
						"type": "boolean",
						"description": "Denotes whether this event represents a failure to perform the action"
					},
					"fields": {
						"$ref": "#/components/schemas/EventFields",
						"description": "An optional set of additional arbitrary event about the data"
					},
					"component": {
						"type": "string",
						"description": "An identifier for the vendor app component that sent the event"
					},
					"version": {
						"type": "string",
						"description": "An identifier for the version of the vendor app that sent the event, usually a git SHA"
					},
					"external_id": {
						"type": "string",
						"description": "An identifier for the environment in which the event was sent"
					},
					"metadata": {
						"$ref": "#/components/schemas/EventFields",
						"description": "An optional set of additional metadata  about the data"
					}
				},
				"required": [
					"action",
					"crud"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"EventNodeGroup": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"FieldItem": {
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventNodeActor": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"href": {
						"type": "string"
					},
					"fields": {
						"items": {
							"$ref": "#/components/schemas/FieldItem"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EventNodeTarget": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"href": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"fields": {
						"items": {
							"$ref": "#/components/schemas/FieldItem"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EventNodeDisplay": {
				"properties": {
					"markdown": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"RawEventNode": {
				"properties": {
					"id": {
						"type": "string"
					},
					"action": {
						"type": "string"
					},
					"crud": {
						"type": "string",
						"enum": [
							"c",
							"r",
							"u",
							"d"
						]
					},
					"description": {
						"type": "string"
					},
					"group": {
						"$ref": "#/components/schemas/EventNodeGroup"
					},
					"actor": {
						"$ref": "#/components/schemas/EventNodeActor"
					},
					"target": {
						"$ref": "#/components/schemas/EventNodeTarget"
					},
					"display": {
						"$ref": "#/components/schemas/EventNodeDisplay"
					},
					"is_failure": {
						"type": "boolean"
					},
					"is_anonymous": {
						"type": "boolean"
					},
					"source_ip": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"loc_subdiv1": {
						"type": "string"
					},
					"loc_subdiv2": {
						"type": "string"
					},
					"received": {
						"type": "string"
					},
					"created": {
						"type": "string"
					},
					"fields": {
						"items": {
							"$ref": "#/components/schemas/FieldItem"
						},
						"type": "array"
					},
					"canonical_time": {
						"type": "string"
					},
					"component": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"external_id": {
						"type": "string"
					},
					"metadata": {
						"items": {
							"$ref": "#/components/schemas/FieldItem"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EventEdge": {
				"properties": {
					"node": {
						"$ref": "#/components/schemas/RawEventNode"
					},
					"cursor": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"PageInfo": {
				"properties": {
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPreviousPage": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"EventsConn": {
				"properties": {
					"edges": {
						"items": {
							"$ref": "#/components/schemas/EventEdge"
						},
						"type": "array"
					},
					"pageInfo": {
						"$ref": "#/components/schemas/PageInfo"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"GraphQLSearch": {
				"properties": {
					"search": {
						"$ref": "#/components/schemas/EventsConn"
					}
				},
				"required": [
					"search"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DocLocation": {
				"properties": {
					"line": {
						"type": "number",
						"format": "double"
					},
					"column": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"line",
					"column"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GraphQLError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"locations": {
						"items": {
							"$ref": "#/components/schemas/DocLocation"
						},
						"type": "array"
					},
					"path": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GraphQLResp": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GraphQLSearch"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/GraphQLError"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"GraphQLSearchVars": {
				"properties": {
					"query": {
						"type": "string",
						"description": "Structured search query. For example,\n\n`edit` -- free text search\n`action:users.list` -- search by action\n`action:document.* location:Germany` -- search by action and location\n`actor.id:john.doe@mycompany.com` -- search by actor id"
					},
					"last": {
						"type": "number",
						"format": "double",
						"description": "The number of events to return"
					},
					"before": {
						"type": "string",
						"description": "An opaque cursor that represents the offset to query since. Should be empty for an initial request to the API. Responses from GraphQL endpoints will include the cursor value to use to fetch the next page of data."
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"GraphQLRequest": {
				"properties": {
					"query": {
						"type": "string",
						"description": "Graphql Search query string. See https://boxyhq.com/docs/retraced/apis/graphql#search\nfor an example query"
					},
					"variables": {
						"$ref": "#/components/schemas/GraphQLSearchVars",
						"description": "Search parameters for filtering, limit and offset"
					},
					"operationName": {
						"type": "string",
						"description": "The GraphQL operation name. Optional, defaults to standard event search."
					}
				},
				"required": [
					"query"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ActiveSearchId": {
				"properties": {
					"id": {
						"type": "string",
						"description": "The ID of the created search"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateActiveSearchRequest": {
				"properties": {
					"saved_search_id": {
						"type": "string",
						"description": "An id of an existing saved search to use"
					}
				},
				"required": [
					"saved_search_id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SavedSearch": {
				"properties": {
					"id": {
						"type": "string",
						"description": "`id` is the id of the saved search. Can be used to initiate an active search"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateSavedSearchRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "`name` of this saved search"
					},
					"actions": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "`actions` is an optional array of action ids to search over"
					},
					"actor_ids": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "`actor_ids` is an optional array of actors to search over"
					},
					"start": {
						"type": "string",
						"format": "date-time",
						"description": "`start` is an optional ISO8601 formatted date to search since"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "retraced",
		"version": "1.9.1",
		"description": "A fully open source audit log service and embeddable UI easily deployed to your own Kubernetes cluster",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
		"/publisher/v1/project/{projectId}/event": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateEventResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "abf053dc4a3042459818833276eec717",
											"hash": "5b570bff4628b35262fb401d2f6c9bb38d29e212f6e0e8ea93445b4e5a253d50"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create an event. Returns the id of the created event, and\na cryptographic hash of the received event, as described at\nhttps://boxyhq.com/docs/retraced/architecture/hashing-formula",
				"tags": [
					"Publisher API"
				],
				"security": [],
				"parameters": [
					{
						"description": "auth header of the form token= ...",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "the project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The event body to log",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateEventRequest",
								"description": "The event body to log"
							}
						}
					}
				}
			}
		},
...
		"/publisher/v1/project/{projectId}/graphql": {
			"post": {
				"operationId": "GraphqlPost",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GraphQLResp"
								}
							}
						}
					}
				},
				"description": "Query events with GraphQL\n\nhttps://boxyhq.com/docs/retraced/apis/graphql",
				"tags": [
					"Publisher API"
				],
				"security": [],
				"parameters": [
					{
						"description": "auth header of the form Token token= ...",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "the project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "graphQL query, variables, and operationName",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GraphQLRequest",
								"description": "graphQL query, variables, and operationName"
							}
						}
					}
				}
			}
		},
		"/enterprise/v1/graphql": {
			"post": {
				"operationId": "GraphqlPost",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GraphQLResp"
								}
							}
						}
					}
				},
				"description": "Query events with GraphQL\n\nhttps://boxyhq.com/docs/retraced/apis/graphql",
				"tags": [
					"Enterprise API"
				],
				"security": [],
				"parameters": [
					{
						"description": "auth header of the form Token token= ...",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "graphQL query, variables, and operationName",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GraphQLRequest",
								"description": "graphQL query, variables, and operationName"
							}
						}
					}
				}
			}
		},
		"/enterprise/v1/search/active": {
			"post": {
				"operationId": "CreateActiveSearch",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActiveSearchId"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "abf053dc4a3042459818833276eec717"
										}
									}
								}
							}
						}
					}
				},
				"description": "Initiate an active search. An active search will maintain\na persistent cursor that can be used at a later date to\nretrieve additional events from the search.\n\n\nAuthenticate with an Enterprise API token.",
				"tags": [
					"Enterprise API"
				],
				"security": [],
				"parameters": [
					{
						"description": "header of the form token= ...",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The search params",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateActiveSearchRequest",
								"description": "The search params"
							}
						}
					}
				}
			}
		},
		"/enterprise/v1/search/saved": {
			"post": {
				"operationId": "CreateSavedSearch",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SavedSearch"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "05abf3dc4a30bf45981883327c7176ee"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a saved search.\nSaved searches have an ID that can be used to initiate an active search.\n\nAuthenticate with an Enterprise API token.",
				"tags": [
					"Enterprise API"
				],
				"security": [],
				"parameters": [
					{
						"description": "header of the form token= ...",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The search parameters",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSavedSearchRequest",
								"description": "The search parameters"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/auditlog",
			"description": "Local"
		}
	]
}